Removed the router file since it prompted an unknown error when starting the app. The GET request was moved to index.ts
